requirement.txt
# API et serveur web
fastapi>=0.95.2
uvicorn>=0.22.0
python-multipart>=0.0.6

# Requêtes HTTP
requests>=2.30.0
httpx>=0.27.0

# Blockchain
xrpl-py>=1.7.0

# Deep Learning
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.2.1
torchvision>=0.17.1
torchaudio>=2.2.1

# Utilitaires
pydantic>=2.7.4
python-dotenv>=1.0.0